+++ tests/test_status.py
@@ -62,7 +62,7 @@ def _common(mocker, tmp_path):
 
     from cpg_flow.targets import MultiCohort
 
-    def mock_create_cohort() -> MultiCohort:
+    def mock_create_cohort(*_) -> MultiCohort:
         m = MultiCohort()
         c = m.create_cohort(id='COH123', name='fewgenomes')
         ds = m.create_dataset('my_dataset')